pyproject.toml
[tool.poetry]
name = "cs496-cloud-computing-project"
version = "0.1.0"
description = ""
authors = ["Ashley Olson <aolson2733@sdsu.edu>", "Adam Larson <alarson7767@sdsu.edu>"]
readme = "README.md"

packages = [
    { include = "devices" }
]

[tool.poetry.dependencies]
python = "^3.11"
azure-iot-device = "^2.14.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
datetime = "^5.5"

[tool.poetry.scripts]
dev = "devices.main:run_dev"
prod = "devices.main:run_prod"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
